python/django/storage/requirements.txt
asgiref==3.6.0
boto3==1.26.118
botocore==1.29.118
Django==4.2
django-storages==1.13.2
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
